+++ crates/nostr-sdk/examples/tor.rs
@@ -15,7 +15,7 @@ async fn main() -> Result<()> {
     let connection: Connection = Connection::new()
         .embedded_tor()
         .target(ConnectionTarget::Onion);
-    let opts = Options::new().connection(connection);
+    let opts = ClientOptions::new().connection(connection);
     let client = Client::builder().signer(keys.clone()).opts(opts).build();
 
     // Add relays